+++ packages/brick_offline_first_build/test/offline_first_model_dictionary_generator_test.dart
@@ -15,7 +15,7 @@ import 'dart:convert';
 import 'package:brick_sqlite/sqlite.dart' show SqliteModel, SqliteAdapter, SqliteModelDictionary, RuntimeSqliteColumnDefinition, SqliteProvider;
 import 'package:brick_rest/rest.dart' show RestProvider, RestModel, RestAdapter, RestModelDictionary;
 // ignore: unused_import, unused_shown_name
-import 'package:sqflite/sqflite.dart' show DatabaseExecutor;
+import 'package:sqflite_common/sqlite_api.dart' show DatabaseExecutor;
 
 import 'person.dart';
 import 'path/user.dart';