+++ packages/brick_offline_first_with_graphql/test/graphql_offline_sqlite_cache_manager_test.dart
@@ -1,7 +1,7 @@
 import 'package:brick_offline_first_with_graphql/src/graphql_offline_queue_link.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache_manager.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql/language.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
@@ -10,7 +10,6 @@ import 'package:sqflite_common/sqlite_api.dart';
 import '__helpers__.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
   sqfliteFfiInit();
 
   group('GraphqlRequestSqliteCacheManager', () {