+++ packages/brick_offline_first_with_rest/test/offline_queue/rest_request_sqlite_cache_test.dart
@@ -1,6 +1,6 @@
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache.dart';
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache_manager.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:http/http.dart' as http;
 import 'package:mockito/mockito.dart';
 import 'package:logging/logging.dart';
@@ -11,7 +11,6 @@ import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 class MockLogger extends Mock implements Logger {}
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
   sqfliteFfiInit();
 
   group('RestRequestSqliteCache', () {