+++ packages/brick_sqlite/lib/sqlite.dart
@@ -1,5 +1,5 @@
-export 'src/sqlite_adapter.dart';
-export 'src/runtime_sqlite_column_definition.dart';
-export 'src/sqlite_model_dictionary.dart';
-export 'src/sqlite_provider.dart';
+export 'package:brick_sqlite/src/sqlite_adapter.dart';
+export 'package:brick_sqlite/src/runtime_sqlite_column_definition.dart';
+export 'package:brick_sqlite/src/sqlite_model_dictionary.dart';
+export 'package:brick_sqlite/src/sqlite_provider.dart';
 export 'package:brick_sqlite_abstract/sqlite_model.dart';