+++ packages/brick_sqlite/test/__mocks__/demo_model_assoc_adapter.dart
@@ -2,7 +2,7 @@ import 'demo_model.dart';
 // ignore: unused_import, unused_shown_name
 import 'package:brick_sqlite_abstract/db.dart';
 import 'package:brick_sqlite/sqlite.dart';
-import 'package:sqflite/sqflite.dart' show DatabaseExecutor;
+import 'package:sqflite_common/sqlite_api.dart' show DatabaseExecutor;
 
 Future<DemoModelAssoc> _$DemoModelAssocFromSqlite(Map<String, dynamic> data,
     {SqliteProvider? provider, repository}) async {