+++ packages/brick_offline_first_with_rest/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Loosen dependency restrictions to major versions
 * Import from `sqflite_common` instead of `sqflite` to avoid a Flutter dependency
 * **BREAKING CHANGE** `OfflineFirstWithRestRepository#offlineQueueHttpClientRequestSqliteCacheManager` is now `offlineQueueManager`
 * **BREAKING_CHANGE** `offlineQueueManager` is required to create `BrickOfflineFirstWithRestRepository`. To migrate without recreating the queue database, pass RestRequestSqliteCacheManager('brick_offline_queue.sqlite', databaseFactory)